# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_deepakgudla_bookvault_internal_dto.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_deepakgudla_bookvault_internal_dto.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  github_com_deepakgudla_bookvault_internal_dto.CreateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
    required:
    - category_id
    - name
    - price
    - sku
    - stock
    type: object
...
  github_com_deepakgudla_bookvault_internal_dto.ProductImageResponse:
    properties:
      alt_text:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      url:
        type: string
    type: object
  github_com_deepakgudla_bookvault_internal_dto.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CategoryResponse'
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductImageResponse'
        type: array
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  github_com_deepakgudla_bookvault_internal_dto.ProductSearchResult:
    properties:
      category:
        $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CategoryResponse'
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductImageResponse'
        type: array
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      rank:
        type: number
      sku:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_deepakgudla_bookvault_internal_dto.UpdateCategoryRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
  github_com_deepakgudla_bookvault_internal_dto.UpdateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - category_id
    - name
    - price
    type: object
...
  github_com_deepakgudla_bookvault_internal_utils.PaginatedResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.PaginationMeta'
      success:
        type: boolean
    type: object
  github_com_deepakgudla_bookvault_internal_utils.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_deepakgudla_bookvault_internal_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:1357
info:
  contact:
    email: deepakgudla35@protonmail.com
    name: Deepak Gudla
    url: https://www.linkedin.com/in/deepakgudla
  description: A modern e-commerce API built with Go
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: bookvault-API
  version: "1.0"
paths:
...
  /api/v1/categories:
    get:
      description: retrieve all categories that are active
      produces:
      - application/json
      responses:
        "200":
          description: categories fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CategoryResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: Get All Categories
      tags:
      - Categories
  /api/v1/categories/:
    post:
      consumes:
      - application/json
      description: create a new product category (Admin Only)
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: successfully created category
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CategoryResponse'
              type: object
        "400":
          description: invalid request data
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Categories
  /api/v1/categories/{id}:
    delete:
      description: delete a category (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: successfully deleted category
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "400":
          description: invalid category ID
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing category (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successfully updated category
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CategoryResponse'
              type: object
        "400":
          description: invalid request data
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Categories
...
  /api/v1/products:
    get:
      description: get the list of active products
      parameters:
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully fetched products
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductResponse'
                  type: array
              type: object
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: Get all products
      tags:
      - Products
  /api/v1/products/:
    post:
      consumes:
      - application/json
      description: create a new product (admin only)
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: successfully created product
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductResponse'
              type: object
        "400":
          description: invalid request data
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Categories
  /api/v1/products/{id}:
    delete:
      description: delete a product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: successfully deleted product
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "400":
          description: invalid product ID
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      description: fetch detailed information about a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully fetched product
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductResponse'
              type: object
        "400":
          description: invalid product ID
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "404":
          description: product not found
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: Get a product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: update an existing product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: update product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successfully fetched product
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductResponse'
              type: object
        "400":
          description: invalid request data
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: update a product
      tags:
      - Products
  /api/v1/products/{id}/images:
    post:
      consumes:
      - multipart/form-data
      description: upload an image for a product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: successfully uploaded image
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: invalid request or file
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Upload product Image
      tags:
      - Products
  /api/v1/search:
    get:
      description: Search Products using full text search with ranking
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      - description: Minimum price filter
        in: query
        name: min_price
        type: number
      - description: Maximum price filter
        in: query
        name: max_price
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductSearchResult'
                  type: array
              type: object
        "400":
          description: Invalid Search query
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Search Products
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: Bearer and JWT token (bearer jwt_token)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"